requirements.txt
alabaster==1.0.0
amqp==5.3.1
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.4
billiard==4.2.1
bleach==6.2.0
celery==5.5.3
celery-progress==0.5
celery-types==0.23.0
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contourpy==1.3.2
crispy-bootstrap4==2025.6
cron-descriptor==1.4.5
cryptography==45.0.5
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
diff-match-patch==20241021
Django==5.2.4
django-celery-beat==2.8.1
django-constance==4.3.3
django-crispy-forms==2.4
django-extensions==4.1
django-import-export==4.3.9
django-maintenance-mode==0.22.0
django-simple-history==3.10.1
django-sslserver==0.22
django-summernote==0.8.20.0
django-tables2==2.7.5
django-timezone-field==7.1
django_celery_results==2.6.0
docopt==0.6.2
docutils==0.21.2
docutils-stubs==0.0.22
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
fonttools==4.59.0
fpdf2==2.8.3
idna==3.10
imagesize==1.4.1
img2pdf==0.6.1
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
josepy==2.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
kombu==5.5.4
ldap3==2.9.1
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mistune==3.1.3
mozilla-django-oidc==4.0.1
mysqlclient==2.2.7
narwhals==2.8.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
num2words==0.5.14
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pdfkit==1.0.0
pexpect==4.9.0
pickleshare==0.7.5
pikepdf==9.10.2
pillow==11.3.0
pipreqs==0.5.0
platformdirs==4.3.8
plotly==6.3.1
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
Pygments==2.19.2
pypandoc==1.15
pyparsing==3.2.3
PyPDF2==3.0.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-fsutil==0.15.0
python3-ldap==0.9.8.4
pytz==2025.2
pyzbar==0.1.9
pyzmq==27.0.0
redis==7.0.1
referencing==0.36.2
requests==2.32.4
roman-numerals-py==3.1.0
rpds-py==0.26.0
scipy==1.16.0
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
snowballstemmer==3.0.1
soupsieve==2.7
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
stack-data==0.6.3
tablib==3.8.0
tinycss2==1.4.0
tornado==6.5.1
traitlets==5.14.3
types-docutils==0.22.2.20250924
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.17.2
xmltodict==0.14.2
yarg==0.1.9
